requirements.txt
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.14.3
google-auth==1.8.2
google-cloud-core==1.1.0
google-cloud-translate==2.0.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
oauthlib==3.1.0
pkg-resources==0.0.0
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
PySocks==1.7.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
tweepy==3.8.0
urllib3==1.25.7
